requirements.txt
blinker==1.9.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.56.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.32.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-nmap==0.7.1
pytz==2025.2
scapy==2.6.1
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
